Kebab Website
My other website I made after the WiJ website.

I changed a lot of things and tried to be a bit more organized with
this one, it’s still a bit messy. The codes that have a giant wall of
code are mostly downloaded from the internet, you shouldn’t have to
waste time writing a jQuery